✨ #159 - SelectProfileImage 컴포넌트 생성 및 적용
